requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
bleach==4.1.0
CacheControl==0.12.10
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
docutils==0.18.1
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
jeepney==0.7.1
keyring==23.3.0
kombu==5.2.2
msgpack==1.0.3
packaging==21.3
pkginfo==1.8.1
prompt-toolkit==3.0.23
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
readme-renderer==30.0
redis==4.0.2
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
twine==3.6.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zipp==3.6.0
